feat: 부모의 액션을 자식에서 처리
